Extract shared Activity logic into a base Activity
